shrink typing
